requirements.txt
absl-py=2.1.0=py311hca03da5_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pypi_0
blas=1.0=openblas
bottleneck=1.4.2=py311hb9f6ed7_0
brotli-python=1.0.9=py311h313beb8_9
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.4.26=py311hca03da5_0
charset-normalizer=3.4.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
cpuonly=2.0=0
debugpy=1.8.14=py311h155a34a_0
decorator=5.2.1=pyhd8ed1ab_0
deprecated=1.2.13=py311hca03da5_0
duckdb=1.2.2=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.1=h313beb8_0
faiss=1.9.0=py311hc0156c1_0_cpu
faiss-cpu=1.9.0=hd610c6a_0
filelock=3.17.0=py311hca03da5_0
frozenlist=1.6.0=pypi_0
fsspec=2024.12.0=py311hca03da5_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py311h5c1b81f_0
grpcio=1.67.1=pypi_0
grpcio-tools=1.71.0=py311h313beb8_0
h11=0.16.0=pypi_0
hf-xet=1.1.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=py311hca03da5_0
ipykernel=6.29.5=pyh57ce528_0
ipython=9.2.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=py311hca03da5_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
krb5=1.21.3=h237132a_0
langchain=0.3.25=pypi_0
langchain-core=0.3.58=pypi_0
langchain-nvidia-ai-endpoints=0.3.9=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langsmith=0.3.42=pypi_0
libabseil=20250127.0=cxx17_h313beb8_0
libcxx=20.1.4=ha82da77_0
libedit=3.1.20191231=hc8eb9b7_2
libfaiss=1.9.0=h682f5e0_0_cpu
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libgrpc=1.71.0=h62f6fdd_0
libmpdec=4.0.0=h80987f9_0
libopenblas=0.3.21=h269037a_0
libprotobuf=5.29.3=h9f9f828_0
libre2-11=2024.07.02=h313beb8_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.45.2=h091b4b1_0
libtorch=2.5.1=gpu_mps_h7b1bc93_203
libuv=1.48.0=h80987f9_0
libzlib=1.2.13=hfb2fe0b_6
llvm-openmp=20.1.3=hdb05f8b_0
markupsafe=3.0.2=py311h80987f9_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
milvus-lite=2.4.12=pypi_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py311hca03da5_0
multidict=6.4.3=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=py311hca03da5_0
nomkl=3.0=0
numexpr=2.10.1=py311h5d9532f_0
numpy=1.26.4=py311he598dae_0
numpy-base=1.26.4=py311hfbfe69c_0
openssl=3.5.0=h81ee809_1
opentelemetry-api=1.30.0=py311hca03da5_0
orjson=3.10.18=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=py311hcf29cfe_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_0
platformdirs=4.3.7=pyh29332c3_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.1=pypi_0
protobuf=5.29.3=py311h514c7bf_0
psutil=7.0.0=py311h917b07b_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pymilvus=2.5.8=pypi_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.0=h3ba56d0_1_cpython
python-dateutil=2.9.0post0=py311hca03da5_2
python-dotenv=1.1.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.11=7_cp311
pytorch=2.5.1=gpu_mps_py311hbd1d365_203
pytorch-mutex=1.0=cpu
pytz=2024.1=py311hca03da5_0
pyyaml=6.0.2=pypi_0
pyzmq=26.4.0=py311h01f2145_0
re2=2024.07.02=h48ca7d4_0
readline=8.2=h1a28f6b_0
regex=2024.11.6=pypi_0
requests=2.32.3=py311hca03da5_1
requests-toolbelt=1.0.0=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
sentence-transformers=4.1.0=pypi_0
setuptools=72.1.0=py311hca03da5_0
six=1.17.0=py311hca03da5_0
sleef=3.5.1=h80987f9_2
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.2=hf2abe2d_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.3=py311hca03da5_1
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.13=h5083fa2_1
tokenizers=0.21.1=pypi_0
tornado=6.4.2=py311h917b07b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.51.3=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.12.2=py311hca03da5_0
tzdata=2025a=h04d1e81_0
ujson=5.10.0=py311h313beb8_1
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py311hca03da5_0
wrapt=1.17.0=py311h80987f9_0
xz=5.6.4=h80987f9_1
yarl=1.20.0=pypi_0
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=py311hca03da5_0
zlib=1.2.13=hfb2fe0b_6
zstandard=0.23.0=pypi_0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytrials==1.0.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0